refactor: Extract chat overlay logic to its own file
