+++ src/index.js
@@ -1,76 +1,12 @@
 require('./tco')
-const { addClass, removeClass, hasClass } = require('./class_utils')
-const createChatContainer = require('./chat_container')
-const createIframe = require('./iframe')
+const { cleanUp, createChatOverlay, toggleChatOverlay } = require('./chat_overlay')
 const createToggle = require('./toggle')
-const { attachBaseStyle, styleToSettings, STYLE_ATTRS } = require('./frame_style')
-const { whenElementLoaded, whenClassToggled, whenUrlChanged, whenKeybindPressed } = require('./observer')
-const { getSettings, setSettings, getGlobalSettings } = require('./settings')
-const makeDraggable = require('./draggable')
-const makeResizable = require('./resizable')
-const { inVOD, getCurrentStream, isTogglingRightColumn, isRightColumnClosed, toggleRightColumn } = require('./current_page')
+const { whenElementLoaded, whenUrlChanged, whenKeybindPressed } = require('./observer')
+const { getSettings, getGlobalSettings } = require('./settings')
+const { inVOD, getCurrentStream } = require('./current_page')
 const setupAutoClaimManager = require('./claim_points')
 
-const enable = _ => addClass(document.body, 'anu-chat-overlay-active')
-
-const disable = _ => removeClass(document.body, 'anu-chat-overlay-active')
-
-let enabled,
-  appendTo,
-  chatContainer,
-  iframe
-
-const createChatOverlay = () => {
-  const appendToParent = document.querySelector('.video-player__overlay')
-  chatContainer = createChatContainer()
-  appendTo = document.createElement('div')
-  iframe = createIframe(_ => {
-    const mouseEventsContainer = document.querySelector('.video-player__overlay')
-    makeResizable(chatContainer, mouseEventsContainer, iframe)
-    makeDraggable(chatContainer, mouseEventsContainer, chatContainer.querySelector('.header'), {
-      onDragEnd: _ => setSettings('position', styleToSettings(chatContainer.style, STYLE_ATTRS.POSITION)),
-      excludedElements: chatContainer.querySelectorAll('.settings, .settings *')
-    })
-
-    attachBaseStyle(iframe.contentDocument.body)
-    iframe.style = ''
-    removeClass(chatContainer, 'loading')
-    const html = document.querySelector('html'),
-      iframeHtml = iframe.contentDocument.querySelector('html'),
-      darkThemeClass = 'tw-root--theme-dark'
-    whenClassToggled(html, darkThemeClass, _ => {
-      if (hasClass(html, darkThemeClass))
-        addClass(iframeHtml, darkThemeClass)
-      else
-        removeClass(iframeHtml, darkThemeClass)
-    })
-  })
-
-  chatContainer.addEventListener('mouseenter', _ => {
-    const chatList = iframe.contentDocument.body.querySelector('.chat-list--default')
-    if (chatList)
-      chatList.scrollTop = chatList.scrollHeight
-  })
-  chatContainer.addEventListener('mouseover', _ => addClass(iframe.contentDocument.body, 'hovered'))
-  chatContainer.addEventListener('mouseout', _ => removeClass(iframe.contentDocument.body, 'hovered'))
-  chatContainer.append(iframe)
-  appendToParent.append(appendTo)
-  appendTo.append(chatContainer)
-}
-
-const toggleChatOverlay = () => {
-  if (!chatContainer)
-    createChatOverlay()
-  enabled = !enabled
-  if (enabled) {
-    enable()
-    if (!isTogglingRightColumn() && window._TCO.currentGlobalSettings.autoCloseRightColumn === 'true' && !isRightColumnClosed()) {
-      toggleRightColumn()
-    }
-  } else {
-    disable()
-  }
-}
+let enabled
 
 const init = async currentStream => {
   if (inVOD())
@@ -84,7 +20,10 @@ const init = async currentStream => {
   setupAutoClaimManager()
 
   const toggle = createToggle()
-  toggle.onclick = toggleChatOverlay
+  toggle.onclick = () => {
+    enabled = !enabled
+    toggleChatOverlay(true, enabled)
+  }
   document.querySelector('.video-player__overlay .player-controls__right-control-group').prepend(toggle)
 
   whenKeybindPressed(() => toggle.click())
@@ -93,26 +32,15 @@ const init = async currentStream => {
   window._TCO.initializing = false
 
   if (enabled) {/* was enabled before the raid/scroll down */
-    createChatOverlay()
+    createChatOverlay(true)
   } else if (window._TCO.currentGlobalSettings.autoStart === 'true') {
     setTimeout(() => toggle.click(), 500);
   }
 }
 
-const cleanUp = _ => {
-  for (const p of document.querySelectorAll('.video-player__overlay .tco-modal'))
-    p.remove()
-  for (const p of document.querySelectorAll('#anu-chat-overlay-toggle'))
-    p.remove()
-  if (appendTo) {
-    appendTo.remove()
-    console.log('Anu Twitch Chat Overlay cleaned up')
-  }
-}
-
 whenElementLoaded(document.body, 'player-controls__right-control-group', async _ => {
   await getGlobalSettings()
-  cleanUp()
+  cleanUp(true)
   await init(getCurrentStream())
 })
 
@@ -122,6 +50,6 @@ whenUrlChanged(async _ => {
     newStream = getCurrentStream()
   if (newStream === oldStream)
     return
-  cleanUp()
+  cleanUp(true)
   await init(newStream)
 }, false)
